content/media/video/test/test_wav_seek3.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Wave Media test: seek test 3</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<audio id='v'
       onloadedmetadata='return startTest();'
       onended='return playbackEnded();'
       onseeking='return seekStarted();'
       onseeked='return seekEnded();'>
  <source type='audio/x-wav' src='r11025_s16_c1.wav'>
</audio>
<pre id="test">
<script class="testbody" type="text/javascript">
// Test seeking works if current time is set before audio is
// playing.
var startPassed = false;
var endPassed = false;
var completed = false;

function startTest() {
  if (completed)
    return false;

  var v = document.getElementById('v');
  v.currentTime=0.5;
  v.play();
  return false;
}

function seekStarted() {
  if (completed)
    return false;

  startPassed = true;
  return false;
}

function seekEnded() {
  if (completed)
    return false;

  endPassed = true;
  return false;
}

function playbackEnded() {
  if (completed)
    return false

  completed = true;
  ok(startPassed, "send seeking event");
  ok(endPassed, "send seeked event");
  SimpleTest.finish();
  return false;
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
